requirements.txt
django
djangorestframework
python-dotenv
earthengine-ap
geemap
geopandas
shapely